+++ core/lib/dal/src/blocks_web3_dal.rs
@@ -5,12 +5,13 @@ use zksync_db_connection::{
 use zksync_system_constants::EMPTY_UNCLES_HASH;
 use zksync_types::{
     api,
+    fee_model::BatchFeeInput,
     l2_to_l1_log::L2ToL1Log,
-    vm_trace::Call,
     web3::{BlockHeader, Bytes},
-    L1BatchNumber, L2BlockNumber, ProtocolVersionId, H160, H2048, H256, U256, U64,
+    Bloom, L1BatchNumber, L2BlockNumber, ProtocolVersionId, H160, H256, U256, U64,
 };
 use zksync_utils::bigdecimal_to_u256;
+use zksync_vm_interface::Call;
 
 use crate::{
     models::{
@@ -43,6 +44,7 @@ impl BlocksWeb3Dal<'_, '_> {
                 miniblocks.timestamp,
                 miniblocks.base_fee_per_gas,
                 miniblocks.gas_limit AS "block_gas_limit?",
+                miniblocks.logs_bloom,
                 prev_miniblock.hash AS "parent_hash?",
                 l1_batches.timestamp AS "l1_batch_timestamp?",
                 transactions.gas_limit AS "transaction_gas_limit?",
@@ -86,7 +88,10 @@ impl BlocksWeb3Dal<'_, '_> {
                         .unwrap_or(i64::from(LEGACY_BLOCK_GAS_LIMIT))
                         as u64)
                         .into(),
-                    // TODO: include logs
+                    logs_bloom: row
+                        .logs_bloom
+                        .map(|b| Bloom::from_slice(&b))
+                        .unwrap_or_default(),
                     ..api::Block::default()
                 }
             });
@@ -174,6 +179,7 @@ impl BlocksWeb3Dal<'_, '_> {
                 miniblocks.timestamp AS "block_timestamp",
                 miniblocks.base_fee_per_gas AS "base_fee_per_gas",
                 miniblocks.gas_limit AS "block_gas_limit?",
+                miniblocks.logs_bloom AS "block_logs_bloom?",
                 transactions.gas_limit AS "transaction_gas_limit?",
                 transactions.refunded_gas AS "transaction_refunded_gas?"
             FROM
@@ -218,7 +224,11 @@ impl BlocksWeb3Dal<'_, '_> {
                         .into(),
                     base_fee_per_gas: Some(bigdecimal_to_u256(row.base_fee_per_gas.clone())),
                     extra_data: Bytes::default(),
-                    logs_bloom: H2048::default(),
+                    logs_bloom: row
+                        .block_logs_bloom
+                        .as_ref()
+                        .map(|b| Bloom::from_slice(b))
+                        .unwrap_or_default(),
                     timestamp: U256::from(row.block_timestamp),
                     difficulty: U256::zero(),
                     mix_hash: None,
@@ -564,17 +574,21 @@ impl BlocksWeb3Dal<'_, '_> {
         .collect())
     }
 
-    /// Returns `base_fee_per_gas` for L2 block range [min(newest_block - block_count + 1, 0), newest_block]
+    /// Returns `base_fee_per_gas` and `fair_pubdata_price` for L2 block range [min(newest_block - block_count + 1, 0), newest_block]
     /// in descending order of L2 block numbers.
     pub async fn get_fee_history(
         &mut self,
         newest_block: L2BlockNumber,
         block_count: u64,
-    ) -> DalResult<Vec<U256>> {
+    ) -> DalResult<(Vec<U256>, Vec<U256>)> {
         let result: Vec<_> = sqlx::query!(
             r#"
             SELECT
-                base_fee_per_gas
+                base_fee_per_gas,
+                l2_fair_gas_price,
+                fair_pubdata_price,
+                protocol_version,
+                l1_gas_price
             FROM
                 miniblocks
             WHERE
@@ -593,10 +607,27 @@ impl BlocksWeb3Dal<'_, '_> {
         .fetch_all(self.storage)
         .await?
         .into_iter()
-        .map(|row| bigdecimal_to_u256(row.base_fee_per_gas))
+        .map(|row| {
+            let fee_input = BatchFeeInput::for_protocol_version(
+                row.protocol_version
+                    .map(|x| (x as u16).try_into().unwrap())
+                    .unwrap_or_else(ProtocolVersionId::last_potentially_undefined),
+                row.l2_fair_gas_price as u64,
+                row.fair_pubdata_price.map(|x| x as u64),
+                row.l1_gas_price as u64,
+            );
+
+            (
+                bigdecimal_to_u256(row.base_fee_per_gas),
+                U256::from(fee_input.fair_pubdata_price()),
+            )
+        })
         .collect();
 
-        Ok(result)
+        let (base_fee_per_gas, effective_pubdata_price): (Vec<U256>, Vec<U256>) =
+            result.into_iter().unzip();
+
+        Ok((base_fee_per_gas, effective_pubdata_price))
     }
 
     pub async fn get_block_details(
@@ -735,9 +766,9 @@ mod tests {
     use zksync_types::{
         aggregated_operations::AggregatedActionType,
         block::{L2BlockHasher, L2BlockHeader},
-        fee::TransactionExecutionMetrics,
         Address, L2BlockNumber, ProtocolVersion, ProtocolVersionId,
     };
+    use zksync_vm_interface::TransactionExecutionMetrics;
 
     use super::*;
     use crate::{