+++ core/lib/multivm/src/versions/vm_virtual_blocks/implementation/gas.rs
@@ -1,7 +1,5 @@
-use zksync_state::WriteStorage;
-
 use crate::{
-    interface::VmInterface,
+    interface::{storage::WriteStorage, VmInterface},
     vm_virtual_blocks::{tracers::DefaultExecutionTracer, vm::Vm},
     HistoryMode,
 };