+++ core/lib/state/src/storage_factory.rs
@@ -6,24 +6,42 @@ use tokio::{runtime::Handle, sync::watch};
 use zksync_dal::{Connection, ConnectionPool, Core, CoreDal};
 use zksync_storage::RocksDB;
 use zksync_types::{L1BatchNumber, StorageKey, StorageValue, H256};
+use zksync_vm_interface::storage::ReadStorage;
 
-use crate::{
-    PostgresStorage, ReadStorage, RocksdbStorage, RocksdbStorageBuilder, StateKeeperColumnFamily,
-};
+use crate::{PostgresStorage, RocksdbStorage, RocksdbStorageBuilder, StateKeeperColumnFamily};
 
-/// Factory that can produce a [`ReadStorage`] implementation on demand.
+/// Storage with a static lifetime that can be sent to Tokio tasks etc.
+pub type OwnedStorage = PgOrRocksdbStorage<'static>;
+
+/// Factory that can produce storage instances on demand. The storage type is encapsulated as a type param
+/// (mostly for testing purposes); the default is [`OwnedStorage`].
 #[async_trait]
-pub trait ReadStorageFactory: Debug + Send + Sync + 'static {
-    /// Creates a [`PgOrRocksdbStorage`] entity over either a Postgres connection or RocksDB
-    /// instance. The specific criteria on which one are left up to the implementation.
+pub trait ReadStorageFactory<S = OwnedStorage>: Debug + Send + Sync + 'static {
+    /// Creates a storage instance, e.g. over a Postgres connection or a RocksDB instance.
+    /// The specific criteria on which one are left up to the implementation.
     ///
-    /// The idea is that in either case this provides a valid [`ReadStorage`] implementation
-    /// that can be used by the caller.
+    /// Implementations may be cancel-aware and return `Ok(None)` iff `stop_receiver` receives
+    /// a stop signal; this is the only case in which `Ok(None)` should be returned.
     async fn access_storage(
         &self,
         stop_receiver: &watch::Receiver<bool>,
         l1_batch_number: L1BatchNumber,
-    ) -> anyhow::Result<Option<PgOrRocksdbStorage<'_>>>;
+    ) -> anyhow::Result<Option<S>>;
+}
+
+/// [`ReadStorageFactory`] producing Postgres-backed storage instances. Hence, it is slower than more advanced
+/// alternatives with RocksDB caches and should be used sparingly (e.g., for testing).
+#[async_trait]
+impl ReadStorageFactory for ConnectionPool<Core> {
+    async fn access_storage(
+        &self,
+        _stop_receiver: &watch::Receiver<bool>,
+        l1_batch_number: L1BatchNumber,
+    ) -> anyhow::Result<Option<OwnedStorage>> {
+        let connection = self.connection().await?;
+        let storage = OwnedStorage::postgres(connection, l1_batch_number).await?;
+        Ok(Some(storage))
+    }
 }
 
 /// DB difference introduced by one batch.
@@ -59,17 +77,17 @@ pub enum PgOrRocksdbStorage<'a> {
     RocksdbWithMemory(RocksdbWithMemory),
 }
 
-impl<'a> PgOrRocksdbStorage<'a> {
-    /// Returns a [`ReadStorage`] implementation backed by Postgres
+impl PgOrRocksdbStorage<'static> {
+    /// Creates a Postgres-based storage. Because of the `'static` lifetime requirement, `connection` must be
+    /// non-transactional.
     ///
     /// # Errors
     ///
-    /// Propagates Postgres errors.
-    pub async fn access_storage_pg(
-        pool: &'a ConnectionPool<Core>,
+    /// Propagates Postgres I/O errors.
+    pub async fn postgres(
+        mut connection: Connection<'static, Core>,
         l1_batch_number: L1BatchNumber,
-    ) -> anyhow::Result<PgOrRocksdbStorage<'a>> {
-        let mut connection = pool.connection().await?;
+    ) -> anyhow::Result<Self> {
         let l2_block_number = if let Some((_, l2_block_number)) = connection
             .blocks_dal()
             .get_l2_block_range_of_l1_batch(l1_batch_number)
@@ -85,9 +103,8 @@ impl<'a> PgOrRocksdbStorage<'a> {
                 .context("Could not find snapshot, no state available")?;
             if snapshot_recovery.l1_batch_number != l1_batch_number {
                 anyhow::bail!(
-                    "Snapshot contains L1 batch #{} while #{} was expected",
-                    snapshot_recovery.l1_batch_number,
-                    l1_batch_number
+                    "Snapshot contains L1 batch #{} while #{l1_batch_number} was expected",
+                    snapshot_recovery.l1_batch_number
                 );
             }
             snapshot_recovery.l2_block_number
@@ -106,12 +123,12 @@ impl<'a> PgOrRocksdbStorage<'a> {
     /// # Errors
     ///
     /// Propagates RocksDB and Postgres errors.
-    pub async fn access_storage_rocksdb(
+    pub async fn rocksdb(
         connection: &mut Connection<'_, Core>,
         rocksdb: RocksDB<StateKeeperColumnFamily>,
         stop_receiver: &watch::Receiver<bool>,
         l1_batch_number: L1BatchNumber,
-    ) -> anyhow::Result<Option<PgOrRocksdbStorage<'a>>> {
+    ) -> anyhow::Result<Option<Self>> {
         tracing::debug!("Catching up RocksDB synchronously");
         let rocksdb_builder = RocksdbStorageBuilder::from_rocksdb(rocksdb);
         let rocksdb = rocksdb_builder