+++ core/lib/vm_interface/src/types/outputs/mod.rs
@@ -1,0 +1,21 @@
+pub use self::{
+    bytecode::CompressedBytecodeInfo,
+    execution_result::{
+        Call, CallType, ExecutionResult, Refunds, TransactionExecutionResult, TxExecutionStatus,
+        VmExecutionLogs, VmExecutionResultAndLogs,
+    },
+    execution_state::{BootloaderMemory, CurrentExecutionState},
+    finished_l1batch::FinishedL1Batch,
+    l2_block::L2Block,
+    statistic::{
+        CircuitStatistic, DeduplicatedWritesMetrics, TransactionExecutionMetrics,
+        VmExecutionMetrics, VmExecutionStatistics, VmMemoryMetrics,
+    },
+};
+
+mod bytecode;
+mod execution_result;
+mod execution_state;
+mod finished_l1batch;
+mod l2_block;
+mod statistic;